requirements.txt
absl-py==1.4.0
acoustics==0.2.6
astunparse==1.6.3
audioflux==0.1.6
audioread==3.0.0
cachetools==5.3.0
flatbuffers==1.12
focal-loss==0.0.7
gast==0.4.0
google-auth==2.16.0
google-pasta==0.2.0
grpcio==1.51.1
imblearn==0.0
joblib==1.2.0
kaldi-io==0.9.4
keras==2.9.0
keras-preprocessing==1.1.2
keras-self-attention==0.51.0
libclang==15.0.6.1
librosa==0.9.2
matplotx==0.3.10
mlxtend==0.23.0
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
pandoc==2.3
pip==24.2
plot==0.6.5
plumbum==1.8.1
ply==3.11
pooch==1.6.0
praat-parselmouth==0.4.3
protobuf==3.19.6
pyaudio==0.2.13
pysef==0.2.10
python-speech-features==0.6
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
smote==0.1
soundfile==0.12.1
termcolor==2.2.0
typing==3.7.4.3
wave==0.0.2
xgboost==1.7.3
